refactor and docs
