+++ python/tvm/micro/base.py
@@ -14,8 +14,7 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-
-"""Base definitions for micro."""
+"""Base definitions for MicroTVM"""
 
 from __future__ import absolute_import
